Add URL format validation
This patch adds validation of the URL structure using the `url`
package and its parsing capability.

Doing this is more complex than one might like, because the `Url` type
does not implement Serde's `Deserialize` trait, so we have to manually
construct a visitor that forwards to `Url::parse` (and which traffics
in the correct error type).

Testing
-------

This patch was tested by passing ill-formed URLs to both the `source`
and `target` parameters.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>