Decode URL using `serde_with`
The `serde_with` package includes a `display_fromstr` (de)serializer
that uses a type’s `Display` and `FromStr` impls to serialize and
deserialize the type, respectively.  Because `Url` has both of these,
we can use `display_fromstr` to deserialize our `Url` without all the
noise of a visitor class.  This patch implements this new
deserialization strategy, at the cost of adding another dependency.

Testing
-------

This patch should not change the observable behavior of the program.
To test this, the same tests that were used to verify the previous
patch were again performed on this patch.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>