Add logic to verify request
This patch adds logic that verifies a Webmention request.  To _verify
a Webmention request_ is to perform a handful of fast checks that
determine whether a Webmention is bogus or unsafe to request.  After
the request is verified, the Webmention itself can be asynchronously
verified.

The following request verification checks are in the Webmention
specification:

  * Check that the source and target both have a supported scheme.  In
    this case, we support `http` and `https` schemes.

  * Check that the source and the target are different.

  * Check that the target points to a supported URL.  For us, we
    support `www.pniedzielski.net` and `pniedzielski.net`.  In the
    future, we may want to make this configurable.

For all of these checks, we perform them without the fragment.

In addition to the above, we perform the following “house”
verification checks:

  * Check that the source is not a loopback address, like `localhost`,
    or the IPv4 and IPv6 loopbacks.  Because we `GET` the source URL,
    we want to avoid being made to access a private resource.

Testing
-------

The changes in this patch were verified manually, checking each of the
above criteria for source and target separately.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>