Log Webmentions received
This patch logs Webmention requests that are well-formed, and whether
they are valid or invalid.  A Webmention request is said to be
_well-formed_ if it has a `source` parameter and a `target` parameter,
both of which are well-formed URLs.  A Webmention request that is
well-formed is said to be _valid_ if it satisfies the validity checks
outlined in the Webmention specification and the “house” validity
checks we implement locally; a Webmention request that is well-formed
but not valid is said to be _invalid_.

For each well-formed Webmention request, this patch causes the program
to log whether the request was valid or invalid, what the source URL
was, and what the target URL was.

Testing
-------

The changes in this patch were manually verified by sending
well-formed valid and invalid Webmentiono requests to the server.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>