Implement asynchronous processing of valid Webmentions
This patch adds support for asynchronously accessing and later
verifying a Webmention given a valid Webmention request.  Right now,
we only implement sending a `GET` request to the source of the
Webmention, without performing any verification of the source page.

Because we are using Actix as our web server, we want to play nicely
with its scheduler.  To do this, we use the Actix spawn capability,
which takes a future and asynchronously schedules it.

Actix comes with an HTTP client, but it doesn’t allow us to follow
redirects, which is required by the specification.  Instead of
manually implementing the correct redirect semantics, we use the
`reqwest` package, which provides an asynchronous HTTP client that
follows redirects.

Testing
-------

The changes in this patch were manually verified by sending valid
Webmention requests to the server, and checking that the server logs a
`GET` request to the source URL, modulo redirects.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>